+++ spec/libraries/install_helpers_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 RSpec.describe TomcatCookbook::InstallHelpers do