+++ test/cookbooks/test/metadata.rb
@@ -4,5 +4,4 @@ license          'Apache-2.0'
 description      'A test cookbook for the Tomcat cookbook'
 version          '0.1.0'
 
-depends 'java'
 depends 'tomcat'